+++ x-pack/plugins/ui_actions_enhanced/public/drilldowns/components/form_drilldown_wizard/form_drilldown_wizard.tsx
@@ -15,7 +15,7 @@ import {
   BaseActionConfig,
   BaseActionFactoryContext,
 } from '../../../dynamic_actions';
-import { ActionWizard } from '../../../components/action_wizard';
+import { ActionWizard } from '../../../components';
 import { Trigger } from '../../../../../../../src/plugins/ui_actions/public';
 import { txtGetMoreActions } from './i18n';
