+++ x-pack/plugins/ui_actions_enhanced/public/drilldowns/drilldown_definition.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import {
+import type {
   ActionFactoryDefinition,
   BaseActionConfig,
   BaseActionFactoryContext,
