+++ x-pack/plugins/ui_actions_enhanced/public/drilldowns/index.ts
@@ -5,5 +5,5 @@
  * 2.0.
  */
 
-export * from './drilldown_definition';
-export * from './components';
+// export * from './drilldown_definition';
+// export * from './components';
